# oas/bow.json
{
  "openapi": "3.0.2",
  "info": {
    "title": "Bow API",
    "description": "Animals event management tool",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:5000",
      "description": "Local environment"
    },
    {
      "url": "https://bow-api-dev.jozuo.work",
      "description": "Development environment"
    },
    {
      "url": "https://bow-api.jozuo.work",
      "description": "Production environment"
    }
  ],
  "paths": {
...
    "/owners/{owner_id}/tasks/": {
      "get": {
        "tags": [
          "tasks"
        ],
        "summary": "タスク情報の一覧取得",
        "description": "オーナーに紐付くタスク情報の一覧を取得します",
        "operationId": "list_owners__owner_id__tasks__get",
        "parameters": [
          {
            "description": "オーナーID",
            "required": true,
            "schema": {
              "title": "Owner Id",
              "pattern": "^[a-z0-9]{32}$",
              "type": "string",
              "description": "オーナーID"
            },
            "name": "owner_id",
            "in": "path"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "title": "Response List Owners  Owner Id  Tasks  Get",
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TaskResponse"
                  }
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "authorization": []
          }
        ]
      },
      "post": {
        "tags": [
          "tasks"
        ],
        "summary": "タスク情報の登録",
        "description": "オーナーに紐付くタスク情報を登録します",
        "operationId": "post_owners__owner_id__tasks__post",
        "parameters": [
          {
            "description": "オーナーID",
            "required": true,
            "schema": {
              "title": "Owner Id",
              "pattern": "^[a-z0-9]{32}$",
              "type": "string",
              "description": "オーナーID"
            },
            "name": "owner_id",
            "in": "path"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TaskRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "title": "Response 400 Post Owners  Owner Id  Tasks  Post",
                  "type": "object"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "authorization": []
          }
        ]
      }
    },
    "/owners/{owner_id}/tasks/{id}": {
      "get": {
        "tags": [
          "tasks"
        ],
        "summary": "タスク情報の1件取得",
        "description": "オーナーに紐付くタスク情報を1件取得します",
        "operationId": "get_owners__owner_id__tasks__id__get",
        "parameters": [
          {
            "description": "オーナーID",
            "required": true,
            "schema": {
              "title": "Owner Id",
              "pattern": "^[a-z0-9]{32}$",
              "type": "string",
              "description": "オーナーID"
            },
            "name": "owner_id",
            "in": "path"
          },
          {
            "description": "タスクID",
            "required": true,
            "schema": {
              "title": "Id",
              "pattern": "^[a-z0-9]{32}$",
              "type": "string",
              "description": "タスクID"
            },
            "name": "id",
            "in": "path"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "title": "Response 404 Get Owners  Owner Id  Tasks  Id  Get",
                  "type": "object"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "authorization": []
          }
        ]
      },
      "put": {
        "tags": [
          "tasks"
        ],
        "summary": "タスク情報の更新",
        "description": "オーナーに紐付くタスク情報を更新します",
        "operationId": "put_owners__owner_id__tasks__id__put",
        "parameters": [
          {
            "description": "オーナーID",
            "required": true,
            "schema": {
              "title": "Owner Id",
              "pattern": "^[a-z0-9]{32}$",
              "type": "string",
              "description": "オーナーID"
            },
            "name": "owner_id",
            "in": "path"
          },
          {
            "description": "タスクID",
            "required": true,
            "schema": {
              "title": "Id",
              "pattern": "^[a-z0-9]{32}$",
              "type": "string",
              "description": "タスクID"
            },
            "name": "id",
            "in": "path"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TaskRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "title": "Response 400 Put Owners  Owner Id  Tasks  Id  Put",
                  "type": "object"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "authorization": []
          }
        ]
      },
      "delete": {
        "tags": [
          "tasks"
        ],
        "summary": "タスク情報の削除",
        "description": "オーナーに紐付くタスク情報を削除します",
        "operationId": "delete_owners__owner_id__tasks__id__delete",
        "parameters": [
          {
            "description": "オーナーID",
            "required": true,
            "schema": {
              "title": "Owner Id",
              "pattern": "^[a-z0-9]{32}$",
              "type": "string",
              "description": "オーナーID"
            },
            "name": "owner_id",
            "in": "path"
          },
          {
            "description": "タスクID",
            "required": true,
            "schema": {
              "title": "Id",
              "pattern": "^[a-z0-9]{32}$",
              "type": "string",
              "description": "タスクID"
            },
            "name": "id",
            "in": "path"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EmptyResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "authorization": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "EmptyResponse": {
        "title": "EmptyResponse",
        "type": "object",
        "properties": {}
      },
...
      "HTTPValidationError": {
        "title": "HTTPValidationError",
        "type": "object",
        "properties": {
          "detail": {
            "title": "Detail",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            }
          }
        }
      },
...
      "TaskRequest": {
        "title": "TaskRequest",
        "required": [
          "icon_no",
          "title",
          "order"
        ],
        "type": "object",
        "properties": {
          "icon_no": {
            "title": "アイコンNo",
            "type": "integer"
          },
          "title": {
            "title": "タイトル",
            "type": "string"
          },
          "order": {
            "title": "画面表示順",
            "minimum": 1,
            "type": "integer"
          },
          "enabled": {
            "title": "有効フラグ",
            "type": "boolean"
          }
        }
      },
      "TaskResponse": {
        "title": "TaskResponse",
        "required": [
          "icon_no",
          "title",
          "order",
          "id",
          "updated_at"
        ],
        "type": "object",
        "properties": {
          "icon_no": {
            "title": "アイコンNo",
            "type": "integer"
          },
          "title": {
            "title": "タイトル",
            "type": "string"
          },
          "order": {
            "title": "画面表示順",
            "minimum": 1,
            "type": "integer"
          },
          "enabled": {
            "title": "有効フラグ",
            "type": "boolean"
          },
          "id": {
            "title": "タスクID",
            "type": "string"
          },
          "updated_at": {
            "title": "更新日時(unixtime)",
            "type": "integer"
          }
        }
      },
      "ValidationError": {
        "title": "ValidationError",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "type": "object",
        "properties": {
          "loc": {
            "title": "Location",
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        }
      }
    },
    "securitySchemes": {
      "authorization": {
        "type": "apiKey",
        "in": "header",
        "name": "x-api-key"
      }
    }
  }
}